Fix logout

+++ frontend/src/Components/LogoutLink/LogoutPopup.js
@@ -15,7 +15,6 @@ function LogoutPopup(props) {
             const token = {
                 refresh_token: localStorage.getItem('refresh_token')
             };
-            let access = localStorage.getItem('access_token');
 
             let res = await axios.post('http://localhost:8000/api/logout/',
                 JSON.stringify(token),
@@ -28,13 +27,17 @@ function LogoutPopup(props) {
                 }
             );
 
-            localStorage.clear();
-            axios.defaults.headers.common['Authorization'] = null;
+            if (res.status == 205) {
+                localStorage.clear();
+                axios.defaults.headers.common['Authorization'] = null;
 
-            props.setTrigger(false)
+                props.setTrigger(false)
 
-            navigate("/");
-            navigate(0);
+                navigate("/");
+                navigate(0);
+            } else {
+                console.log(res);
+            }
 
         } catch (err) {
             console.log(err);
